# api/openapi.yaml
openapi: 3.1.0
info:
  title: Windlass API
  description: |
    HTTP API of the Windlass Docker Compose control plane. All endpoints are
    versioned under /api/v1. Authentication uses an HTTP-only session cookie
    obtained from /auth/login or /auth/setup.
  version: 0.1.0
servers:
  - url: /api/v1
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: windlass_session
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
paths:
...
  /proxy/status:
    get:
      summary: Caddy availability
      operationId: proxyStatus
      security: [{ session: [] }]
      responses:
        "200": { description: "{available, version}" }
...
  /system/docker/images:
    get:
      summary: Docker image disk usage (admin)
      operationId: imageDiskUsage
      security: [{ session: [] }]
      responses:
        "200": { description: Image counts and reclaimable bytes }
  /system/docker/images/prune:
    post:
      summary: Remove old unused images while protecting recent deployments (admin)
      operationId: pruneImages
      security: [{ session: [] }]
      responses:
        "200": { description: Deleted image count and reclaimed bytes }
  /system/panel-domain:
    get:
      summary: Get the Settings-managed panel hostname (admin)
      operationId: getPanelDomain
      security: [{ session: [] }]
      responses:
        "200": { description: Hostname, HTTPS URL, and Caddy availability }
    put:
      summary: Configure or remove the panel hostname in Caddy (admin)
      description: Set hostname to an empty string to remove only the Windlass-owned panel route.
      operationId: setPanelDomain
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hostname]
              properties:
                hostname: { type: string, example: windlass.example.com }
      responses:
        "200": { description: Desired hostname stored and applied }
        "400": { description: Invalid hostname }
...
  /projects:
    get:
      summary: List projects
      operationId: listProjects
      security: [{ session: [] }]
      responses:
        "200": { description: Array of projects }
    post:
      summary: Create a project (member)
      operationId: createProject
      security: [{ session: [] }]
      responses:
        "201": { description: Created with a starter compose.yaml on disk }
        "409": { description: Name already exists }
  /projects/{name}:
    get:
      summary: Get a project
      operationId: getProject
      security: [{ session: [] }]
      responses:
        "200": { description: Project }
        "404": { description: Not found }
    delete:
      summary: Delete a project — containers stopped, directory removed (member; requires password re-entry unless the account is OAuth-only)
      operationId: deleteProject
      security: [{ session: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: The caller's account password. Required when the account has one.
      responses:
        "204": { description: Deleted }
        "403":
          description: Password missing (password_required) or wrong (invalid_password)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /projects/scan:
    post:
      summary: Scan the stacks directory and rebuild the project index (member)
      operationId: scanProjects
      security: [{ session: [] }]
      responses:
        "200": { description: Number of filesystem projects indexed }
  /projects/{name}/files:
    get:
      summary: List project files
      operationId: listProjectFiles
      security: [{ session: [] }]
      responses:
        "200": { description: Array of file infos }
...
  /projects/{name}/env:
    get:
      summary: Read env vars (decrypted)
      operationId: getProjectEnv
      security: [{ session: [] }]
      responses:
        "200": { description: "Object of KEY: value" }
    put:
      summary: Replace env vars (member; stored AES-256-GCM encrypted)
      operationId: setProjectEnv
      security: [{ session: [] }]
      responses:
        "204": { description: Saved }
  /projects/{name}/git:
    put:
      summary: Configure git source (member); returns webhook secret once
      operationId: configureProjectGit
      security: [{ session: [] }]
      responses:
        "200": { description: "{webhook_secret, webhook_url}" }
  /projects/{name}/services:
    get:
      summary: Live service status (compose ps)
      operationId: projectServices
      security: [{ session: [] }]
      responses:
        "200": { description: "{services, note?}" }
  /projects/{name}/logs:
    get:
      summary: Container log stream (SSE)
      operationId: projectLogs
      security: [{ session: [] }]
      parameters:
        - { name: service, in: query, schema: { type: string } }
        - { name: follow, in: query, schema: { type: boolean, default: true } }
        - { name: tail, in: query, schema: { type: integer, default: 200 } }
      responses:
        "200": { description: text/event-stream of log lines }
  /projects/{name}/terminal:
    get:
      summary: Interactive terminal (WebSocket; member)
      description: |
        JSON text frames in ({"type":"input","data":...} / {"type":"resize",...}),
        raw binary terminal output frames out.
      operationId: projectTerminal
      security: [{ session: [] }]
      responses:
        "101": { description: Switching protocols }
  /projects/{name}/actions/{action}:
    post:
      summary: start | stop | restart (member)
      operationId: projectAction
      security: [{ session: [] }]
      responses:
        "204": { description: Done }
        "502": { description: Compose command failed }
  /projects/{name}/deployments:
    get:
      summary: List deployments
      operationId: listDeployments
      security: [{ session: [] }]
      responses:
        "200": { description: Array of deployments, newest first }
    post:
      summary: Deploy (member)
      operationId: createDeployment
      security: [{ session: [] }]
      responses:
        "201": { description: Deployment queued }
        "409": { description: A deployment is already running }
...
  /projects/{name}/domains:
    get:
      summary: List domains with routing status
      operationId: listDomains
      security: [{ session: [] }]
      responses:
        "200": { description: Array of domains }
    post:
      summary: Add a domain (member) — Caddy routes + automatic HTTPS
      operationId: createDomain
      security: [{ session: [] }]
      responses:
        "201": { description: Created }
        "409": { description: Hostname already in use }
...
  /projects/{name}/backups:
    get:
      summary: List backups
      operationId: listBackups
      security: [{ session: [] }]
      responses:
        "200": { description: Array of backups }
    post:
      summary: Back up now (member) — archive + best-effort DB dump
      operationId: createBackup
      security: [{ session: [] }]
      responses:
        "201": { description: Backup record }
...